refactored comments into a new component, then forgot to create a new branch and started on my sort and order functionality
