+++ src/api.js
@@ -4,9 +4,11 @@ const newsAPI = axios.create({
   baseURL: "https://nc-news-backend-project.herokuapp.com/api",
 });
 
-export const fetchArticles = (topic) => {
+export const fetchArticles = (topic, sortBy, order) => {
   return newsAPI
-    .get("/articles", { params: { topic: topic } })
+    .get("/articles", {
+      params: { topic: topic, sort_by: sortBy, order: order },
+    })
     .then(({ data }) => {
       return data.articles;
     });
