+++ src/components/DetailedArticle.jsx
@@ -1,34 +1,16 @@
 import { useParams } from "react-router-dom";
 import { useEffect, useState } from "react";
-import {
-  fetchArticleById,
-  fetchCommentsByArticleId,
-  patchArticleVotesbyId,
-  postCommentbyId,
-  fetchUsers,
-} from "../api";
-import Comment from "./Comment";
-import UserDropDown from "./UserDropDown";
+import { fetchArticleById, patchArticleVotesbyId } from "../api";
+
+import CommentsList from "./CommentsList";
 
 export default function DetailedArticle() {
   const { article_id } = useParams();
 
   const [article, setArticle] = useState({});
   const [isArticleLoading, setIsArticleLoading] = useState(true);
-  const [isCommentsLoading, setIsCommentLoading] = useState(true);
-  const [comments, setComments] = useState([]);
   const [votes, setVotes] = useState(0);
   const [voteErr, setVoteErr] = useState(null);
-  const [postComment, setPostComment] = useState(null);
-  const [postErr, setPostErr] = useState(null);
-  const [isPosting, setIsPosting] = useState(false);
-  const [users, setUsers] = useState([]);
-
-  useEffect(() => {
-    fetchUsers().then((res) => {
-      setUsers(res);
-    });
-  }, []);
 
   useEffect(() => {
     setIsArticleLoading(true);
@@ -38,14 +20,6 @@ export default function DetailedArticle() {
       setIsArticleLoading(false);
     });
   }, [article_id]);
-
-  useEffect(() => {
-    setIsCommentLoading(true);
-    fetchCommentsByArticleId(article_id).then((res) => {
-      setComments(res);
-      setIsCommentLoading(false);
-    });
-  }, [article_id, isPosting]);
 
   const handleUpvote = () => {
     setVotes((currCount) => currCount + 1);
@@ -65,57 +39,10 @@ export default function DetailedArticle() {
     });
   };
 
-  const handlePostComment = () => {
-    setIsPosting(!isPosting);
-    setPostComment(
-      <form onSubmit={handleSubmit} className="post-comment">
-        <div>
-          <label htmlFor="commentbody">Comment:</label>
-          <textarea id="commentbody"></textarea>
-        </div>
-        <div>
-          <UserDropDown users={users} />
-          <input type="submit" value="Submit"></input>
-        </div>
-      </form>
-    );
-  };
-
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    if (event.target[0].value.length <= 10) {
-      setPostErr(
-        <p className="commentErrMsg">comment must be 10 characters or longer</p>
-      );
-    } else if (event.target[1].value === "-------") {
-      setPostErr(<p className="commentErrMsg">please select username</p>);
-    } else {
-      setPostErr(null);
-      postCommentbyId(article_id, {
-        username: event.target[1].value,
-        body: event.target[0].value,
-      }).then((res) => {
-        console.log(res);
-        setPostComment(
-          <Comment
-            key={res.comment_id}
-            comment_id={res.comment_id}
-            votes={res.votes}
-            created_at={res.created_at}
-            author={res.author}
-            body={res.body}
-          />
-        );
-      });
-    }
-  };
-
   const date = new Date(article.created_at);
 
-  console.log("mistake catcher");
-
   if (isArticleLoading) return <p>loading...</p>;
-  if (isCommentsLoading) return <p>loading...</p>;
+
   return (
     <section>
       <h1>{article.title}</h1>
@@ -137,28 +64,10 @@ export default function DetailedArticle() {
           </button>
         </div>
       </div>
-      <div className="commentSection">
-        <h4 className="commentHeader">Comments({article.comment_count}):</h4>
-        <h5 onClick={handlePostComment} className="postComment">
-          Post Comment
-        </h5>
-      </div>
-      <ul className="comments">
-        {postErr}
-        {postComment}
-        {comments.map(({ comment_id, votes, created_at, author, body }) => {
-          return (
-            <Comment
-              key={comment_id}
-              comment_id={comment_id}
-              votes={votes}
-              created_at={created_at}
-              author={author}
-              body={body}
-            />
-          );
-        })}
-      </ul>
+      <CommentsList
+        article_id={article_id}
+        comment_count={article.comment_count}
+      />
     </section>
   );
 }
